Extract format error to avoid dragging winston into shared and renderer
