+++ app/src/lib/logging/main.ts
@@ -1,24 +1,9 @@
 import { getLogger } from '../logging/logger'
+import { formatError } from './format-error'
 
 export interface ILogEntry {
   level: 'info' | 'debug' | 'error'
   readonly message: string
-}
-
-/**
- * Formats an error for log file output. Use this instead of
- * multiple calls to log.error.
- */
-export function formatError(error: Error, title?: string) {
-  if (error.stack) {
-    return title
-      ? `${title}\n${error.stack}`
-      : error.stack.trim()
-  } else {
-    return title
-      ? `${title}\n${error.name}: ${error.message}`
-      : `${error.name}: ${error.message}`
-  }
 }
 
 /**
